chore: simplify tests structure
